use_cases/road_detection/requirements.txt
gradio>=4.0
onnxruntime>=1.16
numpy>=1.24
Pillow>=10.0
rasterio>=1.3
requests>=2.31
shapely>=2.0
geopandas>=1.0
huggingface_hub>=0.20
